What was achieved? Current unit tests now pass against all versions of colander
Bugs Fixed?:
In some versions of colander, integers and floats get cast to strings when the schema was expecting a string.
In retrospect, this will not make the world explode to the end user of the API, so the test checks against a None instead of an integer to keep the test
passing across all versions of colander.

Work Towards User-Stories?:
No? Why Not, create one...